pyproject.toml
[project]
name = "tlwl-workflows"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "notebook>=7.4.7",
    "numpy>=2.3.4",
    "openai>=2.6.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
]
